+++ source/Bestellsystem-Lieferdienst-Client/BL/Datatypes/ProductCategory.cs
@@ -1,0 +1,18 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Bestellsystem_Lieferdienst.BL.Datatypes
+{
+    public class ProductCategory
+    {
+
+        public class ProductCategory
+        {
+            int CategoryID;
+            string CategoryName;
+        }
+    }
+}
