+++ source/Bestellsystem-Lieferdienst-Client/BL/Datatypes/Usertype.cs
@@ -1,0 +1,22 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Bestellsystem_Lieferdienst.BL.Datatypes
+{
+    public class Usertype
+    {
+
+        public struct Usertype
+        {
+            int UsertypeID;
+            string Name;
+        }
+
+
+
+
+    }
+}
